+++ src/interactions/slashCommands/mixins/CommandOptionWithChoices.ts
@@ -1,4 +1,8 @@
-import { APIApplicationCommandOptionChoice, ApplicationCommandOptionType } from 'discord-api-types/v9';
+import {
+	APIApplicationCommandOption,
+	APIApplicationCommandOptionChoice,
+	ApplicationCommandOptionType,
+} from 'discord-api-types/v9';
 import { z } from 'zod';
 import { validateMaxChoicesLength } from '../Assertions';
 import type { ToAPIApplicationCommandOptions } from '../SlashCommandBuilder';
@@ -10,7 +14,9 @@ const choicesPredicate = z.tuple([stringPredicate, z.union([stringPredicate, num
 const booleanPredicate = z.boolean();
 
 export abstract class ApplicationCommandOptionWithChoicesBase<T extends string | number>
-	extends SlashCommandOptionBase
+	extends SlashCommandOptionBase<
+		ApplicationCommandOptionType.String | ApplicationCommandOptionType.Number | ApplicationCommandOptionType.Integer
+	>
 	implements ToAPIApplicationCommandOptions
 {
 	public choices?: APIApplicationCommandOptionChoice[];
@@ -85,10 +91,12 @@ export abstract class ApplicationCommandOptionWithChoicesBase<T extends string |
 			throw new RangeError('Autocomplete and choices are mutually exclusive to each other.');
 		}
 
+		// TODO: Fix types
+		// eslint-disable-next-line @typescript-eslint/consistent-type-assertions
 		return {
 			...super.toJSON(),
 			choices: this.choices,
 			autocomplete: this.autocomplete,
-		};
+		} as APIApplicationCommandOption;
 	}
 }